feat: Location 위치 수정 기능 구현 및 seq Double type으로 변경

+++ src/main/java/com/spot/spotserver/api/schedule/dto/response/LocationResponse.java
@@ -9,7 +9,7 @@ public class LocationResponse {
 
     private Integer day;
 
-    private Integer seq;
+    private Double seq;
 
     private String name;
